// apps/web/src/api/generated/model/userResponseDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */
import type { UserResponseDtoRolesItem } from './userResponseDtoRolesItem';

export interface UserResponseDto {
  id: string;
  username: string;
  /** @nullable */
  email: string | null;
  isSuperAdmin: boolean;
  createdAt: string;
  updatedAt: string;
  employeeUsername?: string;
  permissions?: string[];
  roleIds?: string[];
  /** @nullable */
  avatar?: string | null;
  /**
     * Last successful login timestamp
     * @nullable
     */
  lastLoginAt?: string | null;
  /** Assigned roles with id and name */
  roles?: UserResponseDtoRolesItem[];
}